dl/requirements.txt
numpy
pandas
pyarrow
pyyaml
torch
fastapi
uvicorn
requests
